requirements.txt
torch
torchvision
opencv-python
ultralytics
numpy
matplotlib
albumentations
pytorch-grad-cam
python-dotenv
